requirements.txt
alpaca-py==0.20.3
fastapi==0.111.0
httpx==0.27.0
loguru==0.7.2
numpy==1.26.4
pandas==2.2.2
pandas_market_calendars==4.3.1
plotly==5.22.0
python-dotenv==1.0.1
PyYAML==6.0.2
pydantic==2.7.4
scipy==1.13.1
streamlit==1.36.0
uvicorn==0.30.1
vectorbt==0.26.1
yfinance==0.2.40
pytest==8.2.2
matplotlib==3.9.0

